Protected routes and controllers complete
