+++ api/routes/products.js
@@ -1,7 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const mongoose = require("mongoose");
 const multer = require("multer");
+const checkAuth = require("../middleware/auth-check");
+
+const ProductsControler = require("../controllers/products");
 
 const storage = multer.diskStorage({
   destination: function(req, file, cb) {
@@ -29,154 +31,24 @@ const upload = multer({
   fileFilter: fileFilter
 });
 
-const Product = require("../models/product");
-
 // Get all products from DB
-router.get("/", (req, res, next) => {
-  Product.find()
-    .select("name price _id productImage")
-    .exec()
-    .then(docs => {
-      const response = {
-        count: docs.length,
-        products: docs.map(doc => {
-          return {
-            name: doc.name,
-            price: doc.price,
-            productImage: doc.productImage,
-            _id: doc._id,
-            request: {
-              type: "GET",
-              url: `http://localhost:5000/products/${doc._id}`
-            }
-          };
-        })
-      };
-      res.status(200).json(response);
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.get("/", ProductsControler.product_get_all);
 
 // Add new products to DB
-router.post("/", upload.single("productImage"), (req, res, next) => {
-  console.log(req.file);
-  const product = new Product({
-    _id: new mongoose.Types.ObjectId(),
-    name: req.body.name,
-    price: req.body.price,
-    productImage: req.file.path
-  });
-  product
-    .save()
-    .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: "Created product successfully",
-        createdProduct: {
-          name: result.name,
-          price: result.price,
-          _id: result._id,
-          request: {
-            type: "GET",
-            url: `http://localhost:5000/products/${result._id}`
-          }
-        }
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.post(
+  "/",
+  checkAuth,
+  upload.single("productImage"),
+  ProductsControler.products_create_product
+);
 
 // Get a specific product from DB
-router.get("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  Product.findById(id)
-    .select("name price _id productImage")
-    .exec()
-    .then(doc => {
-      console.log("Data from database: ", doc);
-      if (doc) {
-        res.status(200).json({
-          product: doc,
-          request: {
-            type: "GET",
-            description: "GET_SINGLE_PRODUCT",
-            url: `http://localhost:5000/products/${doc.id}`
-          }
-        });
-      } else {
-        res.status(404).json({
-          message: "No valid data entry for provided ID"
-        });
-      }
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({ error: err });
-    });
-});
+router.get("/:productId", ProductsControler.get_one_product);
 
 // Update a specific object from the DB
-router.patch("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  const updateOps = {};
-  for (const ops of req.body) {
-    updateOps[ops.propName] = ops.value;
-  }
-  Product.update({ _id: id }, { $set: updateOps })
-    .exec()
-    .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: "Product Updated",
-        request: {
-          type: "GET",
-          description: "UPDATE_SINGLE_PRODUCT",
-          url: `http://localhost:5000/products/${id}`
-        }
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.patch("/:productId", checkAuth, ProductsControler.update_one_product);
 
 // Delete a specific object from the DB
-router.delete("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  Product.remove({ _id: id })
-    .exec()
-    .then(result => {
-      res.status(200).json({
-        message: "Product deleted",
-        request: {
-          type: "POST",
-          url: "http://localhost:5000/products",
-          body: {
-            name: "String",
-            price: "Number"
-          }
-        }
-      });
-    })
-    .catch(err => {
-      conosle.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.delete("/:productId", checkAuth, ProductsControler.delete_one_product);
 
 module.exports = router;
